Fix memory leak (causing heap overflow)

+++ packages/strapi/lib/Strapi.js
@@ -202,7 +202,18 @@ class Strapi {
 
   async destroy() {
     this.server.destroy();
-    await this.db.destroy();
+
+    if (_.has(this, 'db')) {
+      await this.db.destroy();
+    }
+
+    if (_.has(this, 'plugins.graphql')) {
+      await this.plugins.graphql.destroy();
+    }
+
+    this.eventHub.removeAllListeners();
+
+    delete global.strapi;
   }
 
   /**
@@ -402,10 +413,10 @@ class Strapi {
     // plugins bootstrap
     const pluginBoostraps = Object.keys(this.plugins).map(plugin => {
       return execBootstrap(_.get(this.plugins[plugin], 'config.functions.bootstrap')).catch(err => {
-        console.log(err);
+        // console.log(err);
         strapi.log.error(`Bootstrap function in plugin "${plugin}" failed`);
         strapi.log.error(err);
-        strapi.stop();
+        // strapi.stop();
       });
     });
     await Promise.all(pluginBoostraps);
@@ -418,7 +429,7 @@ class Strapi {
     return execBootstrap(adminBootstrap).catch(err => {
       strapi.log.error(`Bootstrap function in admin failed`);
       strapi.log.error(err);
-      strapi.stop();
+      // strapi.stop();
     });
   }
